requirements.txt
# FastAPI and ASGI server
fastapi==0.109.2
uvicorn[standard]==0.27.0
gunicorn==21.2.0

# Database
sqlalchemy==2.0.25
alembic==1.13.1
psycopg2-binary==2.9.9
asyncpg==0.29.0
greenlet==3.2.4

# Redis and Celery
redis==5.0.1
celery==5.3.6

# System monitoring
psutil==5.9.8

# Authentication & Security
passlib[bcrypt]==1.7.4
python-jose[cryptography]==3.3.0
python-multipart==0.0.6
cryptography==43.0.3

# Web3 and Blockchain - Using older compatible version
web3==6.11.3
eth-account==0.9.0

# HTTP Client for external APIs
httpx==0.26.0
aiohttp==3.9.3

# Validation and Configuration
pydantic[email]==2.5.3
pydantic-settings==2.1.0
email-validator==2.1.0

# File handling
pillow==10.1.0

# Email and SMS
twilio==8.10.2

# Firebase Cloud Messaging (Push Notifications)
google-auth==2.27.0

# Template rendering
jinja2==3.1.3

# Background tasks and scheduling
flower==2.0.1

# Development and testing
pytest==7.4.4
pytest-asyncio==0.23.4
faker==22.5.1

# Environment and utilities
python-dotenv==1.0.1
pytz==2024.1
